Update README
Note that somewhere along the line from Python 3.6 to 3.10, the number of
errors in the test suite jumped from 2 to 7. (That is, the number of
discrepancies of more than 0.1 between our z score results and the test
values from other implementations.) I did not investigate the cause,
and I'm not entirely positive that the error in this this implementation
(as opposed to the test values themselves being off). They seem to cluster
around the 5-year-old window.